Refactor data table to work with just json
